app/requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-debug
flask-wtf
numpy
pandas
matplotlib
Flask-Script
Flask-Collect